requirements.txt
langchain
langchain-community
streamlit
openai
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
dateparser
